+++ feature-libs/tracking/tms/core/public_api.ts
@@ -1,3 +1,4 @@
 export * from './config/index';
 export * from './model/index';
 export * from './services/index';
+export * from './tms-core.module';
